Verify all "golden" proofs indirectly via reference
