+++ internal/merkle/inmemory/reference_test.go
@@ -25,6 +25,45 @@ import (
 	to "github.com/transparency-dev/merkle/testonly"
 )
 
+// Generated from C++ ReferenceMerklePath, not the Go one so we can verify
+// that they are both producing the same paths in a sanity test.
+var testPaths = []pathTestVector{
+	{0, 1, []string{}},
+	{1, 2, []string{"6e340b9cffb37a989ca544e6bb780a2c78901d3fb33738768511a30617afa01d"}},
+	{
+		0,
+		8,
+		[]string{
+			"96a296d224f285c67bee93c30f8a309157f0daa35dc5b87e410b78630a09cfc7",
+			"5f083f0a1a33ca076a95279832580db3e0ef4584bdff1f54c8a360f50de3031e",
+			"6b47aaf29ee3c2af9af889bc1fb9254dabd31177f16232dd6aab035ca39bf6e4",
+		},
+	},
+	{
+		5,
+		8,
+		[]string{
+			"bc1a0643b12e4d2d7c77918f44e0f4f79a838b6cf9ec5b5c283e1f4d88599e6b",
+			"ca854ea128ed050b41b35ffc1b87b8eb2bde461e9e3b5596ece6b9d5975a0ae0",
+			"d37ee418976dd95753c1c73862b9398fa2a2cf9b4ff0fdfe8b30cd95209614b7",
+		},
+	},
+	{
+		2,
+		3,
+		[]string{"fac54203e7cc696cf0dfcb42c92a1d9dbaf70ad9e621f4bd8d98662f00e3c125"},
+	},
+	{
+		1,
+		5,
+		[]string{
+			"6e340b9cffb37a989ca544e6bb780a2c78901d3fb33738768511a30617afa01d",
+			"5f083f0a1a33ca076a95279832580db3e0ef4584bdff1f54c8a360f50de3031e",
+			"bc1a0643b12e4d2d7c77918f44e0f4f79a838b6cf9ec5b5c283e1f4d88599e6b",
+		},
+	},
+}
+
 // refRootHash returns the root hash of a Merkle tree with the given entries.
 // This is a reference implementation for cross-checking.
 func refRootHash(entries [][]byte, hasher merkle.LogHasher) []byte {
